requirements.txt
# Core dependencies
pydantic>=2.0.0,<3.0.0
aiohttp>=3.8.0,<4.0.0
typing-extensions>=4.5.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0

# For development and testing
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.21.0,<0.22.0
mypy>=1.0.0,<2.0.0
flake8>=6.0.0,<7.0.0
black>=23.0.0,<24.0.0
isort>=5.12.0,<6.0.0
coverage>=7.2.0,<8.0.0
